Use defaults Travis notifications.
From https://docs.travis-ci.com/user/notifications
By default, a build email is sent to the committer and the author, but only if they have access to the repository the commit was pushed to.
The default can be overridden in the .travis.yml. If there’s a setting specified, Travis CI only sends an emails to the addresses specified there, rather than to the committer and author.

[ci skip]